ume_assignment/requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
google-generativeai
gunicorn
drf-yasg